fix tensor type

+++ pyro/distributions/bernoulli.py
@@ -57,8 +57,8 @@ class Bernoulli(Distribution):
             ps = ps.expand_as(x)
         x_1 = x - 1
         ps_1 = ps - 1
-        xmul = torch.mul(x.double(), ps.double())
-        xmul_1 = torch.mul(x_1.double(), ps_1.double())
+        xmul = torch.mul(x.type_as(ps), ps)
+        xmul_1 = torch.mul(x_1.type_as(ps), ps_1)
         logsum = torch.log(torch.add(xmul, xmul_1))
 
         # XXX this allows for the user to mask out certain parts of the score, for example
